lookahead: move connection box to router lookahead map
