+++ vpr/src/base/vpr_context.h
@@ -24,7 +24,6 @@
 #include "place_macro.h"
 #include "compressed_grid.h"
 #include "metadata_storage.h"
-#include "connection_box.h"
 
 //A Context is collection of state relating to a particular part of VPR
 //
@@ -191,8 +190,6 @@ struct DeviceContext : public Context {
     // Name of rrgraph file read (if any).
     // Used to determine when reading rrgraph if file is already loaded.
     std::string read_rr_graph_filename;
-
-    ConnectionBoxes connection_boxes;
 };
 
 //State relating to power analysis
