+++ vpr/src/route/router_lookahead.cpp
@@ -1,7 +1,6 @@
 #include "router_lookahead.h"
 
 #include "router_lookahead_map.h"
-#include "connection_box_lookahead_map.h"
 #include "vpr_error.h"
 #include "globals.h"
 #include "route_timing.h"
@@ -14,8 +13,6 @@ static std::unique_ptr<RouterLookahead> make_router_lookahead_object(e_router_lo
         return std::make_unique<ClassicLookahead>();
     } else if (router_lookahead_type == e_router_lookahead::MAP) {
         return std::make_unique<MapLookahead>();
-    } else if (router_lookahead_type == e_router_lookahead::CONNECTION_BOX_MAP) {
-        return std::make_unique<ConnectionBoxMapLookahead>();
     } else if (router_lookahead_type == e_router_lookahead::NO_OP) {
         return std::make_unique<NoOpLookahead>();
     }
